requirements-laptop.txt
paho-mqtt
python-dotenv
opencv-python
numpy